amend Readme for Seeding instructions
