+++ components/Carousel/Carousel.tsx
@@ -6,13 +6,13 @@ interface CaroProps {
 const Carousel = ({ imagesArray }: CaroProps) => {
   // https://trendyol.github.io/react-carousel/docs/carousel/
   return (
-    <Slider 
-    show={1} 
-    slide={1} 
-    swiping={true}
-    swipeOn={0.05}
-    useArrowKeys={true}
-    transition={0.3}
+    <Slider
+      show={1}
+      slide={1}
+      swiping={true}
+      swipeOn={0.05}
+      useArrowKeys={true}
+      transition={0.3}
     >
       {imagesArray.map((img, index) => {
         return (
