+++ tailwind.config.js
@@ -1,5 +1,5 @@
 module.exports = {
-  mode: 'jit',
+  mode: "jit",
   content: [
     "./pages/**/*.{js,ts,jsx,tsx}",
     "./components/**/*.{js,ts,jsx,tsx}",
@@ -25,4 +25,4 @@ module.exports = {
     extend: {},
   },
   plugins: [],
-}
+};
